+++ meteor/Services/CursorPositionService.cs
@@ -1,13 +1,14 @@
 using System;
+using System.Numerics;
 using meteor.Interfaces;
 
 namespace meteor.Services;
 
 public class CursorPositionService : ICursorPositionService
 {
-    public event Action<int, int[]>? CursorPositionChanged;
+    public event Action<BigInteger, BigInteger[]>? CursorPositionChanged;
 
-    public void UpdateCursorPosition(int position, int[] lineStarts)
+    public void UpdateCursorPosition(BigInteger position, BigInteger[] lineStarts)
     {
         CursorPositionChanged?.Invoke(position, lineStarts);
     }
