+++ meteor/ViewModels/StatusPaneViewModel.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using Avalonia;
 using meteor.Interfaces;
 using ReactiveUI;
@@ -21,14 +22,14 @@ public class StatusPaneViewModel : ViewModelBase
         set => this.RaiseAndSetIfChanged(ref _cursorPosition, value);
     }
 
-    private void OnCursorPositionChanged(int position, int[] lineStarts)
+    private void OnCursorPositionChanged(BigInteger position, BigInteger[] lineStarts)
     {
         int row = 1, column = 1;
         for (var i = 0; i < lineStarts.Length; i++)
         {
             if (position < lineStarts[i]) break;
             row = i + 1;
-            column = position - lineStarts[i] + 1;
+            column = (int)(position - lineStarts[i] + 1);
         }
 
         CursorPosition = new Point(column, row);
